+++ functions.py
@@ -6,6 +6,15 @@ def minus(a, b):
     """Subtract b from a."""
     return a - b
 
+
 def multiply(a, b):
     """multiply function"""
     return a*b
+
+def divide(a,b):
+    """divide function"""
+    if b == 0 :
+        return("나눌수 없음")
+    else:
+        return a/b
+